+++ clash-lib/src/CLaSH/Normalize/Transformations.hs
@@ -897,8 +897,8 @@ reduceConst _ e@(App _ _)
     tcm <- Lens.view tcCache
     reduceConstant <- Lens.view evaluator
     case reduceConstant tcm False e of
-      e'@(Data _)    -> changed e'
       e'@(Literal _) -> changed e'
+      e'@(collectArgs -> (Data _,_)) -> changed e'
       _              -> return e
 
 reduceConst _ e = return e
@@ -937,6 +937,7 @@ reduceConst _ e = return e
 -- * CLaSH.Sized.Vector.unconcat
 -- * CLaSH.Sized.Vector.transpose
 -- * CLaSH.Sized.Vector.replicate
+-- * CLaSH.Sized.Vector.dtfold
 reduceNonRepPrim :: NormRewrite
 reduceNonRepPrim _ e@(App _ _) | (Prim f _, args) <- collectArgs e = do
   tcm <- Lens.view tcCache
@@ -1052,7 +1053,11 @@ reduceNonRepPrim _ e@(App _ _) | (Prim f _, args) <- collectArgs e = do
                then reduceReplicate n aTy eTy vArg
                else return e
           _ -> return e
-
+      "CLaSH.Sized.Vector.dtfold" | length args == 8 ->
+        let ([_kn,_motive,lrFun,brFun,arg],[_mTy,nTy,aTy]) = Either.partitionEithers args
+        in  case runExcept (tyNatSize tcm nTy) of
+          Right n -> reduceDTFold n aTy lrFun brFun arg
+          _ -> return e
       _ -> return e
 
 reduceNonRepPrim _ e = return e